requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
Pillow==9.1.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
